+++ src/commands/reverseAtPointCommands.ts
@@ -1,0 +1,46 @@
+import { window } from 'vscode';
+import { MagitRepository } from '../models/magitRepository';
+import { CommitItemView } from '../views/commits/commitSectionView';
+import { DocumentView } from '../views/general/documentView';
+import { gitRun } from '../utils/gitRawRunner';
+import { StashItemView } from '../views/stashes/stashSectionView';
+import { BranchListingView } from '../views/branches/branchListingView';
+import { RemoteBranchListingView } from '../views/remotes/remoteBranchListingView';
+import { TagListingView } from '../views/tags/tagListingView';
+import { revert } from './revertingCommands';
+import MagitUtils from '../utils/magitUtils';
+
+export async function reverseAtPoint(repository: MagitRepository, currentView: DocumentView): Promise<any> {
+
+  const selectedView = currentView.click(window.activeTextEditor!.selection.active);
+
+  if (selectedView instanceof CommitItemView) {
+
+    const commit = (selectedView as CommitItemView).commit;
+
+    return revert(repository, commit.hash, { noCommit: true });
+
+  } else if (selectedView instanceof BranchListingView ||
+    selectedView instanceof RemoteBranchListingView ||
+    selectedView instanceof TagListingView) {
+
+    const ref = (selectedView as BranchListingView).ref;
+
+    if (ref.commit) {
+      return revert(repository, ref.commit, { noCommit: true });
+    }
+
+  } else if (selectedView instanceof StashItemView) {
+
+    //   const stash = (selectedView as StashItemView).stash;
+
+    //   const args = ['stash', 'apply', '--index', stash.index.toString()];
+    //   return gitRun(repository, args);
+  } else {
+    const ref = await MagitUtils.chooseRef(repository, 'Revert changes', true, true);
+
+    if (ref) {
+      return revert(repository, ref, { noCommit: true });
+    }
+  }
+}