+++ src/views/helpView.ts
@@ -41,7 +41,8 @@ export class HelpView extends DocumentView {
  
 Applying changes
   a Apply          s Stage          u Unstage
-  k Discard        S Stage all      U Unstage all
+  v Reverse        S Stage all      U Unstage all
+  k Discard
   
 Essential commands
   g     refresh current buffer